restructure for typescript conversion
